Renamed multi process sub module
