+++ plaso/multi_process/analysis_process.py
@@ -8,8 +8,8 @@ from plaso.containers import tasks
 from plaso.engine import plaso_queue
 from plaso.lib import definitions
 from plaso.lib import errors
-from plaso.multi_processing import task_process
-from plaso.multi_processing import logger
+from plaso.multi_process import task_process
+from plaso.multi_process import logger
 
 
 class AnalysisProcess(task_process.MultiProcessTaskProcess):
