+++ tests/multi_process/analysis_process.py
@@ -11,10 +11,10 @@ from plaso.containers import sessions
 from plaso.engine import configurations
 from plaso.engine import plaso_queue
 from plaso.engine import zeromq_queue
-from plaso.multi_processing import analysis_process
+from plaso.multi_process import analysis_process
 
 from tests import test_lib as shared_test_lib
-from tests.multi_processing import test_lib
+from tests.multi_process import test_lib
 
 
 class TestAnalysisPlugin(analysis_interface.AnalysisPlugin):
