+++ tests/multi_process/base_process.py
@@ -5,7 +5,7 @@
 import unittest
 
 from plaso.engine import configurations
-from plaso.multi_processing import base_process
+from plaso.multi_process import base_process
 
 from tests import test_lib as shared_test_lib
